merge main features with master

+++ owlinator_app/lib/data/OwlSettings.dart
@@ -5,8 +5,8 @@ class OwlSettings {
   bool mute;
   bool fixedHead;
   bool notify;
-  double angle;
-  double volume;
+  int angle;
+  int volume;
   String assicatedUid;
 
   OwlSettings(this.device, this.mute, this.fixedHead, this.notify, this.angle, this.volume, this.assicatedUid);
@@ -16,8 +16,8 @@ class OwlSettings {
         mute = json['settings']['mute'] as bool,
         fixedHead = json['settings']['fixedHead'] as bool,
         notify = json['settings']['notify'] as bool,
-        angle = (json['settings']['angle'] as double),
-        volume = (json['settings']['volume'] as double),
+        angle = (json['settings']['angle'] as int),
+        volume = (json['settings']['volume'] as int),
         assicatedUid = json['settings']['assicatedUid'] as String;
 
   OwlSettings.defaultSettings(Device device, String uid)
